ggml_cuda_op_rms_norm_fused support fp16

+++ ggml/src/ggml-dlcu/fp16/dl-fp16.cuh
@@ -6,7 +6,7 @@
 void rms_norm_f16_cuda(const half * x, half * dst, const int ncols, const int nrows, const int nchannels, const int nsamples, const int64_t stride_row, const int64_t stride_channel, const int64_t stride_sample, const float eps, cudaStream_t stream);
 
 void rms_norm_mul_f16_cuda(
-    const half * x, const half * mul, half * dst, const int ncols, const int nrows, const int nchannels, const int nsamples,
+    const half * x, const float * mul, half * dst, const int ncols, const int nrows, const int nchannels, const int nsamples,
     const int64_t stride_row, const int64_t stride_channel, const int64_t stride_sample,
     const int64_t mul_stride_row, const int64_t mul_stride_channel, const int64_t mul_stride_sample,
     const int mul_ncols, const int mul_nrows, const int mul_nchannels, const int mul_nsamples,
